// frontend/src/features/center/types/EducationCenterCreateForm.type.ts
export interface EducationCenterCreateForm {
  uuid: string;
  center_name: string;
  center_tel: string;
  ceo_name: string;
  ceo_mobile: string;
  manager_name: string;
  manager_mobile: string;
  center_address: string;
  delivery_address: string;
  unit_price: string;
  center_session: string;
}